+++ compiler-plugin/src/main/java/io/ballerina/stdlib/http/compiler/HttpCompilerPlugin.java
@@ -21,9 +21,13 @@ package io.ballerina.stdlib.http.compiler;
 import io.ballerina.projects.plugins.CompilerPlugin;
 import io.ballerina.projects.plugins.CompilerPluginContext;
 import io.ballerina.projects.plugins.codeaction.CodeAction;
-import io.ballerina.stdlib.http.compiler.codeaction.ChangeHeaderParamTypeToString;
-import io.ballerina.stdlib.http.compiler.codeaction.ChangeHeaderParamTypeToStringArray;
-import io.ballerina.stdlib.http.compiler.codeaction.ChangeReturnTypeWithCaller;
+import io.ballerina.stdlib.http.compiler.codeaction.AddHeaderParameterCodeAction;
+import io.ballerina.stdlib.http.compiler.codeaction.AddPayloadParameterCodeAction;
+import io.ballerina.stdlib.http.compiler.codeaction.AddResponseCacheConfigCodeAction;
+import io.ballerina.stdlib.http.compiler.codeaction.AddResponseContentTypeCodeAction;
+import io.ballerina.stdlib.http.compiler.codeaction.ChangeHeaderParamTypeToStringArrayCodeAction;
+import io.ballerina.stdlib.http.compiler.codeaction.ChangeHeaderParamTypeToStringCodeAction;
+import io.ballerina.stdlib.http.compiler.codeaction.ChangeReturnTypeWithCallerCodeAction;
 
 import java.util.List;
 
@@ -40,9 +44,13 @@ public class HttpCompilerPlugin extends CompilerPlugin {
 
     private List<CodeAction> getCodeActions() {
         return List.of(
-                new ChangeHeaderParamTypeToString(),
-                new ChangeHeaderParamTypeToStringArray(),
-                new ChangeReturnTypeWithCaller()
+                new ChangeHeaderParamTypeToStringCodeAction(),
+                new ChangeHeaderParamTypeToStringArrayCodeAction(),
+                new ChangeReturnTypeWithCallerCodeAction(),
+                new AddPayloadParameterCodeAction(),
+                new AddHeaderParameterCodeAction(),
+                new AddResponseContentTypeCodeAction(),
+                new AddResponseCacheConfigCodeAction()
         );
     }
 }